pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "django-task-manager>=0.1",
    "openai>=1.82.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "ta>=0.11.0",
    "yfinance>=0.2.61",
]
